Listing_E2F_binding_sites_on potato_promoters
The aim of this script is to find a specific transcription factor binding site on the promoters of potato genes. In this case, the E2F transcription factor is used as an example, which binds to sequence variants of the E2F consensus sequence. For this, previously a table (GENE_ID_1500BP_PROMOTER_DICTIONARY.csv) has been created, which contains the gene identification numbers and promoter sequnces of each gene, see Python-pandas-potato-promoter-dictionary-project. Another table was also made before, which contains all the possible sequence variants of the E2F binding sites. With a nested loop, the transcription factor column was iterated through the promoter sequences and with this a combined table was produced that contains the gene id, the particular E2F binding sequence variant and the promoter sequence. In the next step, position of the binding sites is also indenified (indexing) and added to the table. These positions are related to the 3' end of the promoter, which corresponds to the translational start sites.